# swagger.yaml
openapi: 3.1.0
info:
  title: Resi's Project Web Course API
  description: This is Resis's Project Web Course API docs
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
  - url: https://staging.akuanalis.com/api/v1
  - url: https://dev-api.akuanalis.com/api/v1
  - url: https://akuanalis.com/api/v1
...
paths:
...
  /credits:
    get:
      tags:
        - credits
      summary: Get user credit balance
      description: Retrieves the current credit balance for the authenticated user. Credits are earned through course purchases (2 credits per single course, 2 credits per course in a bundle) and consumed by resume analyses (1 credit per analysis).
      operationId: getCredits
      security:
        - userSessionCookie: []
        - adminSessionCookie: []
        - instructorSessionCookie: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditsResponse'
        '401':
          description: Unauthorized - invalid or missing session
        '500':
          description: Internal server error
# end of paths

# start of components
components:
...
  schemas:
...
    CreditsResponse:
      type: object
      properties:
        credits:
          type: integer
          examples:
            - 5

# end of schemas
...
  securitySchemes:
    adminSessionCookie:
      type: apiKey
      in: cookie
      name: session
    userSessionCookie:
      type: apiKey
      in: cookie
      name: session
    instructorSessionCookie:
      type: apiKey
      in: cookie
      name: session